+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 # What's new:
-- Improved lighting on hanging signs chains
-- Added a way to register special renderer through the API (allowing compatibility with mods like Lootr where the block entity is needed to determine the block's appearence)
+- Added partially-incompatible mods detection, in case one is detected, the incompatible settings are turned off and a disclaimer is shown in the config screen